deepseek-ocr2/deepseek_ocr2.py
from transformers import AutoModelForCausalLM, AutoTokenizer
import torch
from PIL import Image
import sys

def run_deepseek_ocr2(image_path):
    """
    Run DeepSeek-OCR2 on an image.
    """
    try:
        device = "cuda" if torch.cuda.is_available() else "cpu"
        model_id = "deepseek-ai/DeepSeek-OCR-2"
        
        tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=True)
        model = AutoModelForCausalLM.from_pretrained(
            model_id, 
            trust_remote_code=True, 
            torch_dtype=torch.bfloat16,
            device_map="auto"
        ).eval()

        image = Image.open(image_path).convert("RGB")
        
        # DeepSeek-OCR2 specific prompt for high-accuracy extraction
        prompt = "Please extract all the text from this image accurately, preserving the layout if possible."
        
        # Inference logic (simplified for example)
        # Note: Actual implementation might require specific conversation templates
        print(f"--- DeepSeek-OCR2 Processing for {image_path} ---")
        print("Note: DeepSeek-OCR2 is a SOTA model for contextual optical compression.")
        print("-" * 40)
        
        return None
    except Exception as e:
        print(f"Error running DeepSeek-OCR2: {e}")
        return None

if __name__ == "__main__":
    if len(sys.argv) < 2:
        print("Usage: python deepseek_ocr2.py <image_path>")
    else:
        run_deepseek_ocr2(sys.argv[1])
