deepseek-ocr2/requirements.txt
transformers
torch
torchvision
Pillow
accelerate
vllm  # Recommended for high-performance inference
